# remove unnecessary FileExists dependency in GetHiFile