It is subsumed by the GetModificationTime dependency.

One less dependency per .hi file,
one less redundant file system access,
five fewer lines of code.

This does *not* implement all the improvements discussed in #583, it's just a minor efficiency fix.

I have not measured the performance impact either, since there is no easy-to-run standard benchmark, but I expect the impact to be very small.