@pepeiborra 2020-05-29
I think it's not a big deal, since the `.hi` file will only be missing the very first time ghcide is started on a given folder. Ever after that, the `.hi` file will exist.