@pepeiborra 2020-05-30
Benchmark BEFORE:
```
TOTAL hover = 2.55s
TOTAL getDefinition = 2.66s
TOTAL documentSymbols = 1.45s
TOTAL documentSymbols after edit = 2.93s
TOTAL completions after edit = 4.55s
TOTAL code actions = 3.59s
Benchmark ghcide-bench: FINISH
```

Benchmark AFTER:
```
TOTAL hover = 2.41s
TOTAL getDefinition = 2.52s
TOTAL documentSymbols = 1.44s
TOTAL documentSymbols after edit = 2.92s
TOTAL completions after edit = 4.42s
TOTAL code actions = 3.52s
Benchmark ghcide-bench: FINISH
```